Add capabilities and server information functions
I think this should be all for RazorNotification. I was thinking of
stripping html tags if there are insufficient capabilities (and so
caching capabilities so that they don't need to be re-queried), but I
don't think it's worth it at this point considering that most (if not
all) daemons support standard tags.