plugin-mount: Don't add devices that are not usable
The disk items shown to the user are the one that are usable
( see  MenuDiskItem::isUsableDevice() ). When a new item is added, it is
shown only if it's an usable one. But when initializing MenuDiskItem
all the devices from the device list are actually added without filtering.
The user can't see them, because they are set to not visible, if they are
not usable. That's a waste of memory.

Signed-off-by: Luís Pereira <luis.artur.pereira@gmail.com>